requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
anyio==4.9.0
attrs==24.2.0
binance==0.3
binance-connector==3.9.0
binance-futures-connector==4.0.1
ccxt==4.4.77
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
contourpy==1.3.0
cryptography==44.0.2
cycler==0.12.1
dateparser==1.2.0
fonttools==4.54.1
frozenlist==1.4.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.2
mplfinance==0.12.10b0
multidict==6.1.0
numpy==1.26.4
packaging==24.1
pandas==2.2.3
pandas_ta==0.3.14b0
pillow==10.4.0
playsound==1.2.2
plyer==2.1.0
pycares==4.6.1
pycparser==2.22
pycryptodome==3.21.0
pyparsing==3.1.4
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2024.2
regex==2024.9.11
requests==2.32.3
scipy==1.15.2
setuptools==78.1.0
six==1.16.0
sniffio==1.3.1
ta==0.11.0
TA-Lib==0.4.32
typing_extensions==4.13.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
websocket-client==1.8.0
websockets==13.1
uvloop==0.21.0
yarl==1.13.1
